requirements.txt
aiofiles>=24.1.0
alive_progress>=3.2.0
appdirs>=1.4.4
art>=6.3
colorama>=0.4.4
fake_useragent>=1.5.1
httpx>=0.27.2
PyYAML>=6.0.2
Requests>=2.32.3
setuptools>=74.1.2
